Added Frontend UI

+++ backend/controllers/goalController.js
@@ -1,12 +1,12 @@
-
 const asyncHandler = require('express-async-handler')
 const Goal = require('../models/goalModel')
+const User = require('../models/userModel')
 
 //@desc Get goals
 //@route GET /api/goals
 //@access Private
 const getGoals = asyncHandler(async (req, res) => {
-  const goals = await Goal.find()
+  const goals = await Goal.find({user: req.user.id})
   res.status(200).json(goals)
 })
 
@@ -20,7 +20,10 @@ const setGoal = asyncHandler(async (req, res) => {
     throw new Error('Please add a text field')
   }
 
-  const goal = await Goal.create({text: req.body.text})
+  const goal = await Goal.create({
+    text: req.body.text, 
+    user: req.user.id,
+  })
 
   res.status(200).json(goal)
 })
@@ -41,6 +44,20 @@ const putGoal = asyncHandler(async (req, res) => {
     throw new Error("Goal not found")
   }
 
+  const user = await User.findById(req.user.id)
+
+  //Check for user
+  if(!user) {
+    res.status(401)
+    throw new Error("User not found")
+  }
+
+  // Make sure the logged in user matches the goal user
+  if(goal.user.toString() !== user.id) {
+    res.status(401) 
+    throw new Error("User not authorized")
+  }
+
   const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {new: true})
 
   res.status(200).json(updatedGoal)
@@ -55,6 +72,20 @@ const deleteGoal = asyncHandler(async (req, res) => {
   if(!goal) {
     res.status(400)
     throw new Error("Goal not found")
+  }
+
+  const user = await User.findById(req.user.id)
+
+  //Check for user
+  if(!user) {
+    res.status(401)
+    throw new Error("User not found")
+  }
+
+  // Make sure the logged in user matches the goal user
+  if(goal.user.toString() !== user.id) {
+    res.status(401) 
+    throw new Error("User not authorized")
   }
 
   const deletedGoal = await Goal.deleteOne(goal)
